Split ui theme module into focused submodules for issue 187
